+++ examples/backend-base/app/config/index.ts
@@ -4,7 +4,6 @@ import startEchoes from './echoes'
 import startGraphQL from './graphql'
 import startHttp from './http'
 import startJobs from './jobs'
-import startTrpc from './trpc'
 import './migrations'
 
 export async function startApp<T extends Component<any>[]>(components: [...T]) {
@@ -14,7 +13,6 @@ export async function startApp<T extends Component<any>[]>(components: [...T]) {
   startGraphQL(controllers.resolvers, controllers.modelResolvers)
   startHttp(controllers.routes)
   startJobs(controllers.jobs)
-  const {router} = startTrpc(controllers.trpc)
 
-  return {router}
+  return {procedures: controllers.trpc}
 }