+++ examples/backend-base/app/exampleComponent/controllers/echoes/GetData/index.ts
@@ -1,10 +1,9 @@
-import {EchoRequest, Echoes} from '@orion-js/echoes'
-import {Inject} from '@orion-js/services'
-import {createEchoRequest} from '@orion-js/echoes'
+import {createEchoRequest, Echoes, EchoRequest} from '@orion-js/echoes'
+import {typedId} from '@orion-js/mongodb'
 import {schemaWithName} from '@orion-js/schema'
+import {Inject} from '@orion-js/services'
 import {ExampleRepository} from 'app/exampleComponent/repos/Example'
 import {ExampleSchema} from 'app/exampleComponent/schemas/ExampleSchema'
-import {typedId} from '@orion-js/mongodb'
 
 const GetDataByIdParams = schemaWithName('GetDataByIdParams', {
   exampleId: {type: typedId('ex')},
@@ -20,6 +19,7 @@ export class GetDataEchoes {
     params: GetDataByIdParams,
     returns: ExampleSchema,
     resolve: async params => {
+      if (!params) throw new Error('Missing params')
       return await this.exampleRepository.getExampleById(params.exampleId)
     },
   })