+++ examples/backend-base/app/exampleComponent/controllers/modelResolvers/ExampleResolvers/index.ts
@@ -1,5 +1,4 @@
-import {ModelResolver, ModelResolvers} from '@orion-js/graphql'
-import {createModelResolver} from '@orion-js/graphql'
+import {createModelResolver, ModelResolver, ModelResolvers} from '@orion-js/graphql'
 import {InferSchemaType} from '@orion-js/schema'
 import {ExampleSchema} from 'app/exampleComponent/schemas/ExampleSchema'
 
@@ -11,6 +10,7 @@ export default class ExampleModelResolvers {
   text = createModelResolver<ExampleType>({
     returns: String,
     resolve: async example => {
+      if (!example) throw new Error('Missing example')
       return `Hello ${example.name}`
     },
   })