+++ examples/backend-base/app/index.ts
@@ -1,6 +1,13 @@
-import { startApp } from "./config";
-import exampleComponent from "./exampleComponent";
+import {logger} from '@orion-js/logger'
+import {registerTRPCRoute} from '@orion-js/trpc'
+import {startApp} from './config'
+import {t} from './config/trpc'
+import exampleComponent from './exampleComponent'
 
-const { router } = await startApp([exampleComponent]);
+const {procedures} = await startApp([exampleComponent])
+const appRouter = t.router(procedures)
 
-export type AppRouter = typeof router;
+registerTRPCRoute({appRouter})
+logger.info('tRPC started at /trpc')
+
+export type AppRouter = typeof appRouter